Fixing minor issues;

+++ client/app/src/main/java/com/qubacy/geoqq/ui/application/activity/_common/screen/geo/settings/GeoSettingsFragment.kt
@@ -143,7 +143,7 @@ class GeoSettingsFragment(
 
     private fun adjustUiWithRadius(radius: Float) {
         mBinding.fragmentGeoSettingsTextRadius.text =
-            getString(R.string.fragment_geo_settings_text_radius_text, radius.toString())
+            getString(R.string.fragment_geo_settings_text_radius_text, radius.toInt().toString())
     }
 
     private fun initPermissionRunner() {
